fix: resolve 10 bugs from iteration 27 bughunt
Backend:
- Replace remaining 9 direct makeError(500,...,getLastError(),...) calls
  in GET handlers with makeDbErrorResponse() — these were still leaking
  raw SQLite internals (column names, constraint names) to clients
- Add RBAC guard before all lab-data write endpoints: VIEWER role can no
  longer create/update/delete samples, orders, or results; API-key auth
  is treated as OPERATOR (minimum write-capable role)
- Fix getSecondsUntilExpiration: return int64_t instead of int to prevent
  signed integer narrowing UB on 64-bit platforms
- Fix PUT /samples, /orders, /results: re-fetch record from DB after
  successful update to return server-committed state (e.g., recalculated
  flag from evaluateFlag()) instead of the pre-commit in-memory object

Frontend:
- AuditLog.tsx handleApplyFilter: was calling fetchAuditLog() which
  closed over the initial mount-time filter state (limit: 50); now
  passes current filter explicitly — filtering by user/action/entity
  was completely broken
- Results.tsx: disable Edit button for VALIDATED results at the list
  level (in addition to the guard inside ResultEditModal) with tooltip
- useEntityList: replace shared cancelledRef with per-effect local
  `cancelled` variable; prevents old request's finally block from
  clearing loading state after a new request has already started

+++ include/auth/JwtAuth.h
@@ -74,7 +74,7 @@ public:
    * @param payload Token payload
    * @return Seconds until expiration (negative if expired)
    */
-  [[nodiscard]] static int getSecondsUntilExpiration(
+  [[nodiscard]] static int64_t getSecondsUntilExpiration(
       const TokenPayload &payload);
 
   /**
